% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumuisha.R
\name{jumuisha}
\alias{jumuisha}
\title{jumuisha}
\usage{
jumuisha(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Summation of x and y
}
\description{
This function takes in two numbers and adds them.
}
\examples{
jumuisha(300, 700)
}
