% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus_35.R
\name{modulus_35}
\alias{modulus_35}
\title{modulus_35}
\usage{
modulus_35(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
A boolean vector where the value is TRUE if an element of the original vector is divisible by 3 or 5, and FALSE otherwise.
}
\description{
Takes in a vector and returns TRUE if an element of the vector is divisible by 3 or 5, and FALSE otherwise.
}
\examples{
modulus_35(c(1:10))

}
